lib/systems/examples: add (x86_64|i686|aarch64)-unknown-windows-msvc
While `x86_64-w64-mingw32abiuefi` might be accepted by `gnuconfig` as
well and may be a better option, it is not easy (?) to compile LLVM for
this target therefore making the whole idea moot.

In the meantime, it uses a LLVM triple which will be compliant to `gnu-config`.

$ ./config.sub x86_64-unknown-windows-msvc
x86_64-unknown-windows-msvc